把播放量改成Long类型，解决追番列表无法显示（凡人播放量超过Int类型的最大值）

+++ app/mobile/src/main/kotlin/dev/aaa1115910/bv/mobile/screen/VideoPlayerScreen.kt
@@ -702,7 +702,7 @@ fun VideoPlayerInfo(
     upFansCount: Int,
     title: String,
     description: String,
-    playCount: Int,
+    playCount: Long,
     danmakuCount: Int,
     date: String,
     avid: Long,
