+++ bili-api/src/main/kotlin/dev/aaa1115910/biliapi/entity/user/Dynamic.kt
@@ -756,7 +756,7 @@ data class DynamicVideo(
     var authorId: Long = 0,
     var authorFace: String = "",
     val duration: Int,
-    val play: Int,
+    val play: Long,
     val danmaku: Int,
     val avatar: String,
     val time: Long = 0L,
@@ -778,7 +778,7 @@ data class DynamicVideo(
                 authorFace = author.face,
                 duration = convertStringTimeToSeconds(archive.durationText),
                 play = convertStringPlayCountToNumberPlayCount(archive.stat.play),
-                danmaku = convertStringPlayCountToNumberPlayCount(archive.stat.danmaku),
+                danmaku = convertStringPlayCountToNumberPlayCount(archive.stat.danmaku).toInt(),
                 avatar = author.face,
                 pubTime = author.pubTime
             )
@@ -808,7 +808,7 @@ data class DynamicVideo(
                         authorFace = author.face,
                         duration = convertStringTimeToSeconds(archive.coverLeftText1),
                         play = convertStringPlayCountToNumberPlayCount(archive.coverLeftText2),
-                        danmaku = convertStringPlayCountToNumberPlayCount(archive.coverLeftText3),
+                        danmaku = convertStringPlayCountToNumberPlayCount(archive.coverLeftText3).toInt(),
                         avatar = author.face
                     )
                 }
@@ -826,7 +826,7 @@ data class DynamicVideo(
                         author = author.name,
                         duration = convertStringTimeToSeconds(pgc.coverLeftText1),
                         play = convertStringPlayCountToNumberPlayCount(pgc.coverLeftText2),
-                        danmaku = convertStringPlayCountToNumberPlayCount(pgc.coverLeftText3),
+                        danmaku = convertStringPlayCountToNumberPlayCount(pgc.coverLeftText3).toInt(),
                         avatar = author.face
                     )
                 }
@@ -842,7 +842,7 @@ data class DynamicVideo(
                         author = author.name,
                         duration = convertStringTimeToSeconds(chargingArchiveInfo.coverLeftText1),
                         play = convertStringPlayCountToNumberPlayCount(chargingArchiveInfo.coverLeftText2),
-                        danmaku = convertStringPlayCountToNumberPlayCount(chargingArchiveInfo.coverLeftText3),
+                        danmaku = convertStringPlayCountToNumberPlayCount(chargingArchiveInfo.coverLeftText3).toInt(),
                         avatar = author.face
                     )
                 }
@@ -868,7 +868,7 @@ private fun convertStringTimeToSeconds(time: String): Int {
 }
 
 //web 接口获取到的是“xx万”，而 grpc 接口获取到的是“xx.x万播放”
-private fun convertStringPlayCountToNumberPlayCount(play: String): Int {
+private fun convertStringPlayCountToNumberPlayCount(play: String): Long {
     if (play.startsWith("-")) return 0
     runCatching {
         val number = play
@@ -876,7 +876,7 @@ private fun convertStringPlayCountToNumberPlayCount(play: String): Int {
             .replace("观看", "")
             .replace("播放", "")
             .substringBefore("万").toFloat()
-        return (if (play.contains("万")) number * 10000 else number).toInt()
+        return (if (play.contains("万")) number * 10000 else number).toLong()
     }.onFailure {
         println("convert play count [$play] failed: ${it.stackTraceToString()}")
     }
