+++ bili-api/src/main/kotlin/dev/aaa1115910/biliapi/http/entity/region/RegionDynamicList.kt
@@ -27,7 +27,7 @@ data class RegionDynamicList(
         val like: Int? = null,
         val name: String,
         val param: String,
-        val play: Int? = null,
+        val play: Long? = null,
         @SerialName("pubdate")
         val pubDate: Int,
         val reply: Int? = null,
